removed unused features, improved cpu load calculation, added socket to transmit gyroscope data, fine-tuned thermistor reading, performance improvements

+++ mobile/app/src/main/java/com/example/esp32_diagtool/model/EspData.kt
@@ -15,5 +15,5 @@ data class EspData(
     @SerializedName("current_total") val currentTotal: Float,
     @SerializedName("battery_life") val batteryLife: String,
     @SerializedName("gpio_pin") val gpioPin: String,
-    @SerializedName("battery_percentage") val batteryPercentage: Int
+    @SerializedName("battery_percentage") val batteryPercentage: Float
 )
